Move user task to scheduler module
