+++ modules/restserver/src/main/scala/docspell/restserver/routes/PeriodicQueryRoutes.scala
@@ -9,7 +9,6 @@ package docspell.restserver.routes
 import cats.data.{NonEmptyList, OptionT}
 import cats.effect._
 import cats.implicits._
-
 import docspell.backend.BackendApp
 import docspell.backend.MailAddressCodec
 import docspell.backend.auth.AuthToken
@@ -20,8 +19,7 @@ import docspell.restapi.model._
 import docspell.restserver.Config
 import docspell.restserver.conv.{Conversions, NonEmptyListSupport}
 import docspell.restserver.http4s.ClientRequestInfo
-import docspell.store.usertask._
-
+import docspell.scheduler.usertask._
 import org.http4s._
 import org.http4s.circe.CirceEntityDecoder._
 import org.http4s.circe.CirceEntityEncoder._
