+++ modules/restserver/src/main/scala/docspell/restserver/routes/ScanMailboxRoutes.scala
@@ -9,14 +9,12 @@ package docspell.restserver.routes
 import cats.data.OptionT
 import cats.effect._
 import cats.implicits._
-
 import docspell.backend.BackendApp
 import docspell.backend.auth.AuthToken
 import docspell.common._
 import docspell.restapi.model._
 import docspell.restserver.conv.Conversions
-import docspell.store.usertask._
-
+import docspell.scheduler.usertask._
 import org.http4s._
 import org.http4s.circe.CirceEntityDecoder._
 import org.http4s.circe.CirceEntityEncoder._
