+++ modules/scheduler/api/src/main/scala/docspell/scheduler/usertask/QUserTask.scala
@@ -1,20 +1,15 @@
-/*
- * Copyright 2020 Eike K. & Contributors
- *
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
+package docspell.scheduler.usertask
 
-package docspell.store.queries
-
-import fs2._
-
+import cats.implicits._
+import cats.effect.Sync
+import com.github.eikek.calev.CalEvent
 import docspell.common._
+import docspell.store.qb.DML
 import docspell.store.qb.DSL._
-import docspell.store.qb._
-import docspell.store.records._
-import docspell.store.usertask.{UserTask, UserTaskScope}
-
-import doobie._
+import docspell.store.records.RPeriodicTask
+import doobie.ConnectionIO
+import fs2.Stream
+import io.circe.Encoder
 
 object QUserTask {
   private val RT = RPeriodicTask.T
@@ -100,5 +95,65 @@ object QUserTask {
 
   def makeUserTask(r: RPeriodicTask): UserTask[String] =
     UserTask(r.id, r.task, r.enabled, r.timer, r.summary, r.args)
+
+  def create[F[_]: Sync](
+      enabled: Boolean,
+      scope: UserTaskScope,
+      task: Ident,
+      args: String,
+      subject: String,
+      priority: Priority,
+      timer: CalEvent,
+      summary: Option[String]
+  ): F[RPeriodicTask] =
+    Ident
+      .randomId[F]
+      .flatMap(id =>
+        Timestamp
+          .current[F]
+          .map { now =>
+            RPeriodicTask(
+              id,
+              enabled,
+              task,
+              scope.collective,
+              args,
+              subject,
+              scope.fold(_.user, identity),
+              priority,
+              None,
+              None,
+              timer,
+              timer
+                .nextElapse(now.atZone(Timestamp.UTC))
+                .map(_.toInstant)
+                .map(Timestamp.apply)
+                .getOrElse(Timestamp.Epoch),
+              now,
+              summary
+            )
+          }
+      )
+
+  def createJson[F[_]: Sync, A](
+      enabled: Boolean,
+      scope: UserTaskScope,
+      task: Ident,
+      args: A,
+      subject: String,
+      priority: Priority,
+      timer: CalEvent,
+      summary: Option[String]
+  )(implicit E: Encoder[A]): F[RPeriodicTask] =
+    create[F](
+      enabled,
+      scope,
+      task,
+      E(args).noSpaces,
+      subject,
+      priority,
+      timer,
+      summary
+    )
 
 }
