+++ modules/scheduler/api/src/main/scala/docspell/scheduler/usertask/UserTaskScope.scala
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-package docspell.store.usertask
+package docspell.scheduler.usertask
 
 import docspell.common._
